[add] 스토어 리스트 컴포넌트 생성
