+++ src/components/store/storeList/StoreList/index.ts
@@ -1,0 +1,1 @@
+export * from './StoreList';
